app/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inngest==0.5.11
itsdangerous==2.2.0
jcs==0.2.1
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.7.2
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
redis==5.2.1
